+++ km/creds/kube/testdata/kube_ca.crt
